refactor(lsp): improve package structure
