+++ quarkdown-lsp/src/main/kotlin/com/quarkdown/lsp/completion/function/impl/name/FunctionNameCompletionSupplier.kt
@@ -1,8 +1,9 @@
-package com.quarkdown.lsp.completion
+package com.quarkdown.lsp.completion.function.impl.name
 
 import com.quarkdown.lsp.cache.CacheableFunctionCatalogue
 import com.quarkdown.lsp.cache.DocumentedFunction
-import com.quarkdown.quarkdoc.reader.DocsWalker
+import com.quarkdown.lsp.completion.CompletionSupplier
+import com.quarkdown.lsp.completion.function.FunctionCallInsertionSnippet
 import org.eclipse.lsp4j.CompletionItem
 import org.eclipse.lsp4j.CompletionItemKind
 import org.eclipse.lsp4j.CompletionParams
@@ -12,14 +13,14 @@ import java.io.File
 
 /**
  * Provides completion items for function names in function calls by scanning documentation files.
- * This supplier is proxied by [FunctionCompletionSupplier] and expects already-sliced text.
+ * This supplier is proxied by [com.quarkdown.lsp.completion.function.FunctionCompletionSupplier] and expects already-sliced text.
  * @param docsDirectory the directory containing the documentation files to extract function data from
  */
 internal class FunctionNameCompletionSupplier(
     private val docsDirectory: File,
 ) : CompletionSupplier {
     /**
-     * Converts a [DocsWalker.Result] to a [CompletionItem] for use in function name completion.
+     * Converts a [com.quarkdown.quarkdoc.reader.DocsWalker.Result] to a [org.eclipse.lsp4j.CompletionItem] for use in function name completion.
      */
     private fun toCompletionItem(function: DocumentedFunction) =
         CompletionItem().apply {
