+++ quarkdown-lsp/src/main/kotlin/com/quarkdown/lsp/completion/function/impl/parameter/FunctionParameterAllowedValuesCompletionSupplier.kt
@@ -1,4 +1,4 @@
-package com.quarkdown.lsp.completion
+package com.quarkdown.lsp.completion.function.impl.parameter
 
 import com.quarkdown.lsp.cache.DocumentedFunction
 import com.quarkdown.lsp.tokenizer.FunctionCall
