+++ quarkdown-lsp/src/main/kotlin/com/quarkdown/lsp/hover/HoverSuppliersFactory.kt
@@ -1,10 +1,11 @@
 package com.quarkdown.lsp.hover
 
 import com.quarkdown.lsp.QuarkdownLanguageServer
+import com.quarkdown.lsp.hover.function.FunctionDocumentationHoverSupplier
 
 /**
  * Factory for creating a list of [HoverSupplier]s.
- * @property server the Quarkdown language server instance
+ * @param server the Quarkdown language server instance
  */
 class HoverSuppliersFactory(
     private val server: QuarkdownLanguageServer,
@@ -14,6 +15,8 @@ class HoverSuppliersFactory(
      */
     fun default() =
         listOf(
-            FunctionDocumentationHoverSupplier(docsDirectory = server.docsDirectory ?: error("Docs directory not available")),
+            FunctionDocumentationHoverSupplier(
+                docsDirectory = server.docsDirectory ?: error("Docs directory not available"),
+            ),
         )
 }
